Use distinct symbols for raw and founder genotypes
Constrain genotype values so that raw alleles must be represented as a single positive digit in the range 1-9, founder alleles must be represented by a single uppercase or lowercase letter, and genotypes must be represented by a concatenation of P alleles, where P is the sample ploidy. This will enable automatic detection of genotype category (raw or founder), as well as genotype ploidy.