Update estimation and output of thresholds
Add a new `scanonebins` class for estimation of empirical false-discovery rates (FDRs), as well as functions to create, manipulate, and display `scanonebins` objects. Allow creation of 'scanonebins' in batchPermScanone. Add functions to read and write `scanonebins` objects in HDF5 format. Change default HDF5 result names to allow for different types of scanone permutation data.

Add option to set FDR in `getQTLPeaks` and `getQTLIntervals`, as well as the `scanone` pipeline. Reform `plotQTLIntervals` to `plotQTLScanone`; update to allow for output of very dense markers, and add the option of displaying an FDR threshold.

Change `report` pipeline to use updated plot function, and to use the (more unicode-friendly) Cairo PDF device.

Add utility functions `emptyArgs`, `isSingleFiniteNumber`, and `hasEnumGenotypes`.

Add a new dependency - the `abind` package - and update `README.md` file to reflect this.