Add additional checks to validateGenotypeSet
Check genotype symbols are sorted and unique, and do not contain blanks.

Check that genotypes are of consistent ploidy. This requirement should be relaxed if there is a cross that may have different ploidy between or within individuals.

Check that genotypes are haploid and have two genotypes. These last two requirements should be relaxed if appropriate as new cross types are added.